Clean up / slight simplification
